Addressbuch-WIP-LW18-Unterricht, tried to get the maven dependencies to work, to be able to debug the javax.awt & swing GUI, but still WIP
