--- IF-LW15-D-GUI-Addressbuch/Addressbuch/AddressbuchGUI.jfm
@@ -1,192 +1,0 @@
-object AddressbuchGUI: TFGUIForm
-  Left = 1171
-  Top = 328
-  BorderIcons = [biSystemMenu]
-  Caption = 'Addressbuch'
-  ClientHeight = 334
-  ClientWidth = 215
-  Color = clBtnFace
-  ParentFont = True
-  FormStyle = fsStayOnTop
-  OldCreateOrder = True
-  Position = poDesigned
-  ShowHint = True
-  Visible = True
-  OnClose = FormClose
-  OnCloseQuery = FormCloseQuery
-  OnMouseActivate = FormMouseActivate
-  OnResize = FormResize
-  FrameType = 5
-  Modified = False
-  Resizable = False
-  Undecorated = False
-  Background = clBtnFace
-  PixelsPerInch = 96
-  TextHeight = 15
-  object jTextField1: TJTextField
-    Tag = 2
-    Left = 24
-    Top = 16
-    Width = 158
-    Height = 20
-    Cursor = crIBeam
-    Hint = 'jTextField1'
-    Font.Charset = DEFAULT_CHARSET
-    Font.Color = clWindowText
-    Font.Height = -12
-    Font.Name = 'Dialog'
-    Font.Style = []
-    Foreground = 3355443
-    Background = clWhite
-    HorizontalAlignment = LEFT
-    Text = ''
-    Editable = True
-  end
-  object jNumberField1: TJNumberField
-    Tag = 21
-    Left = 24
-    Top = 80
-    Width = 155
-    Height = 20
-    Cursor = crIBeam
-    Hint = 'jNumberField1'
-    Font.Charset = DEFAULT_CHARSET
-    Font.Color = clWindowText
-    Font.Height = -12
-    Font.Name = 'Dialog'
-    Font.Style = []
-    Foreground = 3355443
-    Background = clWhite
-    HorizontalAlignment = LEFT
-    Text = ''
-    Editable = True
-  end
-  object jButton1: TJButton
-    Tag = 4
-    Left = 24
-    Top = 48
-    Width = 75
-    Height = 25
-    Hint = 'jButton1'
-    Font.Charset = DEFAULT_CHARSET
-    Font.Color = clWindowText
-    Font.Height = -12
-    Font.Name = 'Dialog'
-    Font.Style = [fsBold]
-    Foreground = 3355443
-    Background = 15658734
-    actionPerformed = 'jButton1_ActionPerformed'
-    Text = 'jButton1'
-    Mnemonic = 0
-    DisplayedMnemonicIndex = 0
-    Selected = False
-    BorderPainted = True
-    FocusPainted = False
-    ContentAreaFilled = True
-    HorizontalAlignment = CENTER
-    VerticalAlignment = CENTER
-    HorizontalTextPosition = RIGHT
-    VerticalTextPosition = CENTER
-    IconTextGap = 4
-    RolloverEnabled = False
-    Border.BorderType = NoBorder
-    Border.LineColor = clBlack
-    Border.LineThickness = 0
-    Border.LineRounded = False
-    Border.EtchHighlightColor = clBlack
-    Border.EtchShadowColor = clBlack
-    Border.Etchtype = 0
-    Border.BevelHighlightColor = clBlack
-    Border.BevelShadowColor = clBlack
-    Border.Beveltype = 0
-    Border.MatteColor = clBlack
-    Border.MatteTop = 0
-    Border.MatteLeft = 0
-    Border.MatteBottom = 0
-    Border.MatteRight = 0
-  end
-  object jLabel1: TJLabel
-    Tag = 1
-    Left = 24
-    Top = 112
-    Width = 150
-    Height = 44
-    Hint = 'jLabel1'
-    Font.Charset = DEFAULT_CHARSET
-    Font.Color = clWindowText
-    Font.Height = -12
-    Font.Name = 'Dialog'
-    Font.Style = [fsBold]
-    Foreground = 3355443
-    Background = clBtnFace
-    Text = 'text'
-    HorizontalAlignment = LEFT
-    VerticalAlignment = CENTER
-    HorizontalTextPosition = RIGHT
-    VerticalTextPosition = CENTER
-    IconTextGap = 4
-    DisplayedMnemonic = 0
-    DisplayedMnemonicIndex = 0
-  end
-  object jComboBox1: TJComboBox
-    Tag = 9
-    Left = 24
-    Top = 176
-    Width = 150
-    Height = 20
-    Hint = 'jComboBox1'
-    Font.Charset = DEFAULT_CHARSET
-    Font.Color = clWindowText
-    Font.Height = -12
-    Font.Name = 'Dialog'
-    Font.Style = [fsBold]
-    Foreground = 3355443
-    Background = 15658734
-    Editable = False
-    SelectedIndex = -1
-  end
-  object jButton2: TJButton
-    Tag = 4
-    Left = 104
-    Top = 48
-    Width = 75
-    Height = 25
-    Hint = 'jButton2'
-    Font.Charset = DEFAULT_CHARSET
-    Font.Color = clWindowText
-    Font.Height = -12
-    Font.Name = 'Dialog'
-    Font.Style = [fsBold]
-    Foreground = 3355443
-    Background = 15658734
-    actionPerformed = 'jButton2_ActionPerformed'
-    Text = 'jButton2'
-    Mnemonic = 0
-    DisplayedMnemonicIndex = 0
-    Selected = False
-    BorderPainted = True
-    FocusPainted = False
-    ContentAreaFilled = True
-    HorizontalAlignment = CENTER
-    VerticalAlignment = CENTER
-    HorizontalTextPosition = RIGHT
-    VerticalTextPosition = CENTER
-    IconTextGap = 4
-    RolloverEnabled = False
-    Border.BorderType = NoBorder
-    Border.LineColor = clBlack
-    Border.LineThickness = 0
-    Border.LineRounded = False
-    Border.EtchHighlightColor = clBlack
-    Border.EtchShadowColor = clBlack
-    Border.Etchtype = 0
-    Border.BevelHighlightColor = clBlack
-    Border.BevelShadowColor = clBlack
-    Border.Beveltype = 0
-    Border.MatteColor = clBlack
-    Border.MatteTop = 0
-    Border.MatteLeft = 0
-    Border.MatteBottom = 0
-    Border.MatteRight = 0
-  end
-end
