+++ IF-LW15-D-GUI-Addressbuch/Kontakt.java
